requirements.txt
streamlit
google-generativeai
pymupdf
docx
dotenv

